refactored bins and data batches into a subpackage
